requirements.txt
# Cat Project - Python dependencies
# For PC (video generation):
Pillow>=10.0
opencv-python>=4.8
numpy>=1.24

# For Raspberry Pi (playback + WhatsApp bot):
flask>=3.0
twilio>=9.0
